+++ src/org/gearvrf/bullet/shape/ConeShape.java
@@ -20,7 +20,7 @@ import org.gearvrf.bullet.util.ShapeType;
 public class ConeShape implements Shape {
 	
 	private static final int type = ShapeType.CONE_SHAPE_PROXYTYPE;
-	private int id;
+	private long id;
 
 	public final float radius;
 	public final float height;
@@ -36,12 +36,12 @@ public class ConeShape implements Shape {
 	}
 
 	@Override
-	public int getID() {
+	public long getID() {
 		return id;
 	}
 	
 	@Override
-	public void setID(int id) {
+	public void setID(long id) {
 		this.id = id;
 	}
 }
